refactor(duelling-picklist): move focus to item buttons
