+++ src/components/duelling-picklist/index.d.ts
@@ -1,5 +1,5 @@
-export { default as DuellingPicklist } from './duelling-picklist';
-export { default as PicklistDivider } from './picklist-divider';
-export { default as PicklistItem } from './picklist-item';
-export { default as Picklist } from './picklist';
-export { default as PicklistPlaceholder } from './picklist-placeholder';
+export { default as DuellingPicklist } from "./duelling-picklist";
+export { default as PicklistDivider } from "./picklist-divider/picklist-divider";
+export { default as PicklistItem } from "./picklist-item/picklist-item";
+export { default as Picklist } from "./picklist/picklist";
+export { default as PicklistPlaceholder } from "./picklist-placeholder/picklist-placeholder";
