+++ src/components/duelling-picklist/index.js
@@ -1,5 +1,5 @@
 export { default as DuellingPicklist } from "./duelling-picklist.component";
-export { default as PicklistDivider } from "./picklist-divider.component";
-export { default as PicklistItem } from "./picklist-item.component";
-export { default as Picklist } from "./picklist.component";
-export { default as PicklistPlaceholder } from "./picklist-placeholder.component";
+export { default as PicklistDivider } from "./picklist-divider/picklist-divider.component";
+export { default as PicklistItem } from "./picklist-item/picklist-item.component";
+export { default as Picklist } from "./picklist/picklist.component";
+export { default as PicklistPlaceholder } from "./picklist-placeholder/picklist-placeholder.component";
