+++ src/components/duelling-picklist/picklist-divider/picklist-divider.component.js
@@ -1,6 +1,6 @@
 import React from "react";
 
-import { StyledPicklistDivider } from "./duelling-picklist.style";
+import StyledPicklistDivider from "./picklist-divider.style";
 
 const PicklistDivider = (props) => (
   <StyledPicklistDivider {...props} data-element="picklist-divider" />
