+++ src/components/duelling-picklist/picklist-divider/picklist-divider.style.js
@@ -1,0 +1,14 @@
+import styled from "styled-components";
+
+const StyledPicklistDivider = styled.div`
+  min-width: 2px;
+  background-image: linear-gradient(
+    180deg,
+    #bfcbd1 0%,
+    rgba(191, 203, 209, 0) 99.9%
+  );
+  margin-left: 4px;
+  margin-right: 4px;
+`;
+
+export default StyledPicklistDivider;
